exynos3: New PowerHAL
Kanged from https://github.com/OMAP4-AOSP/android_device_samsung_tuna/commit/5c61bce12475b0387c38492ed6418912be563799 and name and initial parameters changed.
Supports OC, limits cpufreq to 800Mhz when screen is off, CPU boot when POWER_HINT_INTERACTIVE, only supports ondemand governor

Bit of history on the previous PowerHAL - it was kanged from d2 (SGS3 in North America which has even less in common with us than the Galaxy Nexus).  Was designed primarily for the ondemand governor but the interactive governor (which it also supported) is used now.  Appeared to not function correctly either, ie when the screen was on it set things as if it were off and vice versa.  exynos3 has never had a PowerHAL designed for it as its too old.